requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
crispy-bootstrap4==2023.1
Django==4.2.2
django-crispy-forms==2.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
mysql-connector-python==8.0.22
phonenumbers==8.12.21
Pillow==8.2.0
protobuf==3.17.3
PyMySQL==1.0.2
pytz==2020.5
six==1.16.0
sqlparse==0.4.1
typing-extensions==4.6.3
tzdata==2023.3
